coffee-backend/requirements.txt
# 커피 취향 기반 AI 추천 + 감성 기록 웹앱 requirements

# FastAPI (백엔드 API 서버, 필요시)
fastapi==0.110.2
uvicorn[standard]==0.29.0

# Firebase Admin SDK (Firestore 연동)
firebase-admin==6.4.0
google-cloud-firestore==2.13.0

# OpenAI API
openai==1.30.1

# LangGraph (멀티에이전트)
langgraph==0.0.38

# Kakao Map API, 크롤러, 파서
requests==2.31.0
beautifulsoup4==4.12.3

# 기타
python-dotenv==1.0.1

# (선택) 데이터 분석/취향 요약용
pandas==2.2.2

# (선택) 챗봇 플로우 제어
langchain==0.1.16

# (선택) 웹 프론트엔드 연동용
# streamlit==1.35.0 